pyproject.toml
[project]
name = "BackEnd"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "renato-chagas", email = "renatochagas.m@gmail.com"},
]
dependencies = ["django>=5.2.8", "djangorestframework>=3.16.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm.scripts]
dev = "python src/manage.py runserver"
createsuperuser = "python src/manage.py createsuperuser"
pre_migrate = "python src/manage.py makemigrations"
migrate = "python src/manage.py migrate"
post_migrate = "python src/manage.py graph_models -g -o core.png core"
shell = "python src/manage.py shell"